generate_lab_grade_reports/requirements.txt
google-api-python-client>=2.92.0
google-auth>=2.22.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
